Removed DeprecatedCache from automock
